xtts_service/requirements.txt
# --- SERVIDOR Y API ---
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
typing-extensions>=4.8.0

# --- MOTOR DE INTELIGENCIA ARTIFICIAL (VOZ) ---
TTS==0.22.0
torch==2.2.1
transformers==4.37.2
spacy==3.7.4
huggingface-hub==0.25.0

# --- EDICIÓN DE AUDIO ---
pydub==0.25.1
# Quitamos el pin fijo y dejamos el mínimo que pide TTS
scipy>=1.11.2